Adds portfolio tab and react-router
adds portfolio and filter. Filter needs to be able to filter projects by language.
It will do this by having a map of the languages and true or false of if it is selected
if it is selected the filter should show it lit up and the projects relating to it will be shown